Adding AbstractBitbucketTeam to move shared code
